create storyCard component
